put GuardOrderGenerator in the correct namespace
